+++ src/MyConnection.h
@@ -43,6 +43,20 @@ public:
     );
   }
 
+  std::string quoteString(std::string input) {
+    // Create buffer with enough room to escape every character
+    std::string output(' ', input.length() * 2 + 1);
+
+    mysql_real_escape_string(pConn_, &output[0], input.c_str(), input.size());
+
+    // Find null
+    size_t end = output.find_first_of('\0');
+    if (end != std::string::npos)
+      output.resize(end);
+
+    return output;
+  }
+
   ~MyConnection() {
     try {
       mysql_close(pConn_);